xdg-shell: create and implement xdg_surface_present()
xdg_surface_present() is a new request, supposed to be
called on an existing xdg_surface, and which will let
the shell inform the user about this particuler surface.
The user may eventually raise & focus the surface if he
reacts accordingly.

We create a "managed_surface" interface in desktop-shell,
to keep track of shell surfaces in the weston-desktop-shell
client if something needs to be drawn later.

Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>