desktop-shell: implement presented() tooltip notifier
Each time a managed_surface received a "presented"
event, a red tooltip will be created on the panel and
stacked on the right side (near the clock).

The user can click on the tooltip to make the shell
send an "activate" request and make the surface raise
to the foreground of the current workspace.

The tooltip is then destroyed. It is also destroyed if the
corresponding shell surface gets destroyed by any means.

Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>