xdg-shell: implement a serial argument for xdg_surface_present()
If a serial is given as an additional argument to the present()
request, and is different of 0 (0 is unlikely because serials are global
and increment for each event), we look if it matches a previous serial
issued for a wl_pointer, wl_touch or wl_keyboard event.
If it does, we look if this event is not older than 2 seconds ; if it
is, issue a tooltip notification, if is not, raise and focus the
surface.

Change-Id: If0b54b5c0507918c37b56f885aeec1ef98ad36c1
Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>